ComicSite no longer extends Site
This was a BIG hassle in terms of db migration.. hence 6 migrations.. Lots
of errors with missing UNIQUE contraints, duplicate primary keys etc.. But
I guess a good training for complex south migration.
Running these migrations will not allow going back, the ComicSite.id will
be changed and previous value not saved. But I think this is no problem