Fixed chunked uploading. Files chunks now are associated with a single file.
Added support for chunking on the server-side. Chunks now get a client-defined id for
re-associating chunks on the serverside. Model for StagedFiles changes significantly to
support this. Chunked files are stored as fragments and will be reassembled later by the
dedicated AjaxUpload-form field.

Changed the nginx-configuration to oonly allow small uploads. Big uploads now are chunked.