TRAC#7301 Improves single upload widget
The form is now submitted when a single file has completed uploading.
This should be ok as we do not need any more info. Multifileuploads
still work but don't look great, this is ok as we're not going to use
them right now - instead, use modal forms to get the submission,
description, etc, etc.