TRAC#7384 Caches the Result absolute url
The get_absolute_url method makes a single database query, so this takes
a long time on the results list view. Therefore, cache the url and
store it on the object. This will cache until the next Result object
save, so will everything will need to be updated if the ComicSite
short_name changes, or the results object url changes.