Add unfiltered page option
Challenge administrators are considered trusted users, so when they
create pages they have access to the Source button for ckeditor. This
allows a further option where the Source is not filtered client side
which can be enabled on a per-challenge basis. This is disabled by
default so that the look and feel of new challenges fits in to the
existing website.

Fix #442