Change upload_path to upload_path_hash, add client-side generated uploadsessions ids
The client-side upload session ids are only used to check for duplicate files against
the server's database. They are not security relevant, since this is a UI-tweak allowing
the server to reject file uploads when before they are actually happening. This should
save bandwidth and also gives quicker user-responsed in case a user tries to upload
a duplicate file name.

Issue #479