Fix top level redirect
The urls on the top level were not correctly implementing APPEND_SLASH,
it seems that there is a well known bug in django about this. This
commit swaps out the catch all url for the FlatpageFallbackMiddleware
which correctly handles a 404 page with APPEND_SLASH=True.

https://code.djangoproject.com/ticket/6213

As reverse lookups for flatpages no longer work we have to construct
the full url from the request.site, but, we can now remove a workaround
in the url resolver. A side effect of this now is that the flatpages
will appear on the subdomains, but the pages on those domains have
precedence.

Closes #957