Handle non 502 error codes
When workstation sessions are started concurrently it can
take a while for traefik to route them, resulting in some
running workstations returning 404.

We use Math.Random to add jitter to the clients who are
trying to connect. A max_attempts parameter is added so
that the client will try for ~30 seconds (currently)
after the workstation is started to connect to it,
before giving up.

Closes #1143