Account for ground truth in progress, export and when adding answers (#1150)
* Always calculate completed hangings

Checking of the remainder of answer_count / answerable_question_count is zero could potentially mark too many hangings as completed. E.g. if there are 10 hangings with one image each and 10 questions, answering five questions per hanging would result in a progress of 50% where it should be 0%.

* Cache leaderboard and statistics

They get accessed quite often in the template. This prevents the db queries from being executed each time `.statistics` or `.leaderboard`  is called in the templates, which is for at least once for each entry.