Enhance websocket proxy capture (#1360)
Current situation: Cirrus workstation does not use MWT options 'webSocketBasePath'. Instead it makes a call to /Settings/webSocketPort, which not only returns the port 443 but also the path needed for GC to correctly proxy the ws request.
SATORI on the other hand uses 'webSocketBasePath' and sets it to the path of window.location. This leads to a URI that includes the workstation initial path. But the nginx location to match 'mlab4d4c4142' URI does not match then.
Solution: cirrus also makes use of 'webSocketBasePath' and the nginx location adds a regex capture to strip everything between the workstation hostname and mlab4d4c4142.

This PR implements the later