Fix worker giving up on broker loss
Sets broker_connection_max_retries to 0 so that the celery workers
do not stop trying to reconnect to a lost broker. This prevents
the celery workers having to be restarted when the broker is down
for more than 100s.

See https://docs.celeryproject.org/en/stable/userguide/configuration.html#broker-connection-max-retries