Give challenge admins permissions to view algorithm evaluations
This allows challenge admins to view the outputs to the algorithm evaluations. It does
not allow them to run the algorithm on new data, even with legacy submission as they
do not have permission to use the algorithm. This does not leak the test set to the
participant or algorithm creators as only the challenge admins can view this info.