Fix downloading of image files in development (#1677)
This PR allows developers to download image files from minio-protected.localhost,
which Firefox allows mixed HTTP and HTTPS traffic from. Also removes setting
the hosts in the developers host file as the urls now work natively with
Firefox 84 and Edge (not tested in Chrome but I believe this is ok).