Add average successful job duration to algorithm api
Adds the average successful job duration for this algorithm to the api.
If there are no successful jobs, returns None/null.
If there are successful jobs, the serialized value is a float of the duration in seconds.
The average duration will be updated everytime a new successful job has completed for that algorithm.