Use copy rather than move
We should not modify the source directory.
The TIFF strategy needs re-thinking a little. First, create the image objects
from tiff in the output directory if they are valid, then, for all output tiff
files try and create a corresponding dzi. Skipping this right now as I do
not have enough test data.