Fix social media cards for subpages (#1768)
* Fix social media cards for subpages

Object is only present for the archive, challenge, algorithm or reader study
detail pages. This information will be missing for subpages of that object.
Use firstof to get the containing object, allowing us to override this later
by setting a `social_info` object if we need something specific.

* Fix meta headers

firstof only returns a string, so split this out to work with objects.