Allow auth token to last longer than the session (#1780)
Sometimes it takes a few minutes for the system to kill a session,
which means that the auth token will be invalid whilst the session
is still usable. A minute grace period allows the user to finish
up during which time the container will be stopped by the
"grandchallenge.components.tasks.stop_expired_services" periodic task.