Add accelerated client (#2105)
Transfer acceleration is not supported by the VPC gateway, so if we use transfer
acceleration for all of our interactions then everything goes through the NAT
Gateway. Only the end users need to use the accelerated endpoint as our compute
is in the same region as the bucket, so only enable transfer acceleration when
generating the presigned urls for the end user.