Make license detection dynamic
This changes things around so that open source determination
is done at run time, rather that at import time. This allows
us to modify the list of licenses considered open source.

Licensee is upgraded to v9.15.1 so that json output is available.
This standardises the results and opens up more options such
as dynamically determining fine-grained restrictions such as
commercial use.