Don't cache is_editable as aggressively (#2360)
The aggressice caching leads to issues when checking if a display set can be edited, as the display set's modified time does not get updated when an answer is created. Updating the display set when this happens seems counter productive, as that would also invalidate th cache for the value list, slowing down the view.