Gather widgets media in htmx list view (#2503)
`htmx` requires that all of the forms media is loaded first.
If we are using a form that is rendered by htmx in a list view,
then we need to know about all of the possible forms and their
media up front. The widgets used in a form are generated dynamically,
so for each form used in a htmx view we declare up front what
widgets are possible, and gather their media.

An error is logged to sentry if a form is rendered with an
unexpected widget.