Fix admin task to requeue jobs
For some reason doing queryset.update then iterating through
the queryset to execute the jobs didn't work. This instead
does a bulk update and requeing of the job one by one and
appears to fix the issue. We no longer need to issue the command
twice to requeue a job.