Fix json file validation (#2711)
The schema validation in `validate_user_upload` didn't work because the
file contents weren't parsed correctly. This function is only called in
two places (ds file updating and adding a file to a civ). The tests for
both scenarios were incomplete and didn't catch the issue. I updated the
tests accordingly.

Closes #2713