Check if variables are not None to allow values of zero (#2723)
Fixes
https://sentry.io/organizations/grand-challenge/issues/3814067033/?project=303639&query=is%3Aunresolved&referrer=issue-stream

There should always be a budget now for requests. For this particular
case, the budget ended up being `None` because they had filled in 0s for
the phase 2 specs, and so the conditional `if self.phase_2_... `
evaluated to False, when it really should not (setting 0 should be
allowed, definitely for Type 2 specs but also for the other fields - the
budget stuff usually needs to be adjusted and updated again after
discussion with the organizers anyway).