Refactor algorithm job creation (#2780)
This PR refactors the entrypoints for algorithm job creation. It removes
the session upload over the HTML forms functionality as it is only
possible to use for a subset of algorithms with the single GMI input.
Batch upload for all algorithms is available over the REST API instead.

The naming of views and forms are updated to reflect that they are only
creating one job. References to "experiment" are removed. Earlier
validation is added.

I could have gone further and turned the `JobCreateForm` to a
`ModelForm` so that all of the logic is there rather than on the view
(would make more sense) but that is too much work for now.

Closes #1903 
See https://github.com/DIAGNijmegen/rse-roadmap/issues/207