Separate out shuffling tests (#2791)
This PR separates out the test of the shuffling from the large
shuffling, indexing and answering test. What was happening is that the
seed would be set for the database such that the shuffled order would be
the same as the original order, so checking that `shuffled_order !=
original_order` failed. The new test creates two users with different
pks and different seeds, so they should receive different orders. To
increase the likelihood I have increased the number of display sets
used. I don't know if the 2nd part of the test still works if the
ordering is unchanged, but hopefully that is OK as the test is run often
with different seeds. I have run each test 60 times and not managed to
reproduce the problem, so hopefully this fixes the flaky test.

Closes #2759