Add a Number Input Widget for Reader Studies (#2816)
This PR adds a `NumberInput` widget for reader study questions. This new
widget can only be selected for answers of type `Number`. The user can
optionally configure `min`, `max` and `step size` values for this
widget, which if defined get validated for answers for this questions.

Addresses the GC side of
https://github.com/DIAGNijmegen/rse-roadmap/issues/220

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>