Add is_desired_version to ComponentImage model (#2819)
- This adds a non-editable field `is_desired_version` to the
`ComponentImage` model. When a new image is created and its manifest has
been validated, this field is set to True.
- This also adds a method to mark an image as the desired version, which
takes care of updating all other images associated to the same algorithm
/ phase / workstation. This method is wrapped in a transaction, so that
either all images are updated or none.
- Adds a management command to set `is_desired_version` to `True` for
all currently active images after running the migrations (which will set
this field to False for all existing images).

The new field is currently not used to determine which image to use for
an algorithm run etc. It seemed best to first make sure that all images
have this field set appropriately before making the remaining changes
(i.e. to `latest_executable_image`).

This is the first part of
https://github.com/DIAGNijmegen/rse-roadmap/issues/223