Fix missing source maps (#2837)
Django 4.0 a change to
[ManifestStaticFilesStorage](https://docs.djangoproject.com/en/4.2/ref/contrib/staticfiles/#django.contrib.staticfiles.storage.ManifestStaticFilesStorage)
means that Django tries to hash the JS source maps. Missing maps have
been added but these are still often missing from distributions, so
before we run `collectstatic` empty source maps are created that fixes
the choking of `ManifestStaticFilesStorage`. Really a bit of a hack,
but, I couldn't find a better way. The maps will still not be available,
but this is OK.