Introduce TEXT-type answers (#2873)
Part of the pitch:
- https://github.com/DIAGNijmegen/rse-roadmap/issues/165

This change-set adds TEXT-type answers including:
* widgets `Text Input`, `Text Area`
* validation params `answer_max_length`, `answer_min_length`,
`answer_match_pattern`

Changeset also includes:
* a minor bug in the `NUMBER` + `NUMBER_INPUT` widget that showed when
adding an `answer_type` factor to the parameterized tests
* fix not showing the help-text of both the `interface` and the `widget`
fields in the form (the `widgets` + `help_text` + `DynamicField` do not
seem to work as one would expect.)

It purposely does **not** contain the data migration to collapse the
to-be deprecated 'Multi line' and 'Single line' answer types. This
requires a stepped release in that the previous release of CIRRUS should
support TEXT types.