Add Direct Messages (#3093)
This PR adds the Direct Messages application. It introduces three main
new models:

- `Conversation`: Many users are `participants` of one conversation.
Many permissions are added to each conversation - view,
create_conversation_direct_message, mark_conversation_read,
mark_conversation_message_as_spam.
- `DirectMessage`: Has a foreign key to a `conversation`, `sender` and
which users have not read this message. The sender can delete the
message and the conversation participants can mark the message as spam.
If a direct message is created the other participants in the group have
the message marked as unread. If a message is deleted or marked as spam
then all of the unread markers are deleted.
- `Mute`: tracks which users have muted which other users. If a direct
message is created and the receiver has muted the sender the message is
not marked as unread by the receiver.

The views use HTMX. As a conversation is loaded all of the messages are
marked as read as a callback. For now only challenge admins can message
their participants.

See https://github.com/DIAGNijmegen/rse-roadmap/issues/255