My challlenge requests list (#3099)
This updates the challenge requests list view so that both reviewers and
users can access it and view the requests they have access to. To do
this, this adds object permissions for challenge requests. All related
views are updated with new permissions and a reviewer group is added.

Once this is merged, we will need to add the 10% team members to the new
reviewer group, as well as Chris, Miriam, me and James.

Closes #3060