Send email on permissions change (#3122)
This emails the admins of the challenge if a Phase is turned into an
open one. Just so that they're really aware. I also added a method to
send through a standard email as we have a lot of places that do this in
different ways.

See https://github.com/DIAGNijmegen/rse-roadmap/issues/257