Remove the legacy submission creation form (#3133)
This was only really useful for type 1 challenges (which are no longer a
focus) as the challenge are usually not editors of the participants
algorithms, and it complicates the submission limit tracking. Creating
such submissions is possible for the editor, they can then ask us to
re-assign the user if needs be.