Streamline CIV creation and validation code for archive items and display sets (#3156)
Closes #3138 
Part of https://github.com/DIAGNijmegen/rse-roadmap/issues/274 

With this change to use the same logic and code for CIV creation for
display sets and archive items, it is no longer possible to update the
interface of an image in an archive item. We already discussed this
[here ](https://github.com/comic/grand-challenge.org/pull/3116)briefly
and had decided that this was no longer necessary.