Fix json editor initialization (#3163)
Closes #3162 

Fix for the issue was to call `htmx.onLoad` only after all DOM content
is loaded. For the display set views to work then, the
`search_for_jsoneditor_widgets` needs to be called outside of the event
listener as well though. Finally, I'm making sure that in case the
function gets called multiple times, we only initialize the widget once
and only for the relevant element (if it is triggered by an htmx load
event).