Add intended use (#3184)
A paper is has been accepted and proofs sent in which we analyze the
intended use statements of AI products. In the paper we promise to have
them available on AIforRadiology.com. The statements were collected
almost two years ago. I want to give vendors the chance to double check
if they are still up to date. Therefore in this version they are
somewhat 'hidden'. When the paper comes live, I will submit a pull
request to change the template to show it more clearly to all visitors.