Set current value on flexible image widget (#3225)
This makes sure that any existing value is passed to the
FlexibleImageWidget. The widget now looks as follows when there is an
existing value:

![image](https://github.com/comic/grand-challenge.org/assets/30069334/21270e80-35c5-4a44-aa94-6729b58a4bba)

vs when there is none:

![image](https://github.com/comic/grand-challenge.org/assets/30069334/40ed7b62-66ba-43dc-8148-38ce2c6c9249)
 


I've added some comments so that future devs know why sometimes
`initial` is a CIV and sometimes not. I also considered moving the file
type widget logic from `MultipleCIVForm` to the `InterfaceFormField`,
but decided against it because the `SelectUploadWidget` that we use for
file interfaces with an initial value requires a bunch of extra
information that is specific to the `MultipleCIVForm` and that seemed
odd to pass on to `InterfaceFormField`. The `SelectUploadWidget` is also
only used in the MultipleCIVForm, so seemed ok to me to keep the logic
there.

Closes #3205