Add a YouTube tag for MarkDown fields (#3100)
This PR adds a markdown notation `[youtube ID]` to embed youtube into
markdown on pages. These can be created by users, require `<iframe>`s
and are shown to users.

This markdown:
```Markdown
# A Title
This is a demo reader study.

## Instructions
Just look at how wonderful this loads. Note that it's set to load things on a 16:9, hence the best result is gotten by recording on that aspect ratio.

[youtube QCYYhkTlnhQ]
```

![image](https://github.com/comic/grand-challenge.org/assets/7871310/8a2513be-f9a6-4ee8-ac71-ef524651fc1d)

## Custom Tags
A post-processing step to parse new custom tags is now made. This is
done after the bleaching process and operates on clean HTML. As
such, a custom tag should produce clean HTML itself.

It should be pretty straight forward to extend this with new tags.

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>