Move `DockerExecutor` to tests (#3273)
This PR moves the `DockerExecutor` to the tests app so that it is not
present in the production containers and renames it to
`InsecureDockerExecutor`. It also adds a deployment check to make sure
users are not using this in production, and raises
`ImproperlyConfigured` if `DEBUG=False` and `pytest` is not being run.
If `DEBUG=True` we add a critical error message on every initialisation.
This should prevent users from using it in production.