Simplify json error reporting (#3286)
Json validation errors used to include the instance (the submitted
json), the schema and the error. This changes the reporting to only
print the error message without the schema and instance. This is a
temporary fix until this is implemented:
https://github.com/python-jsonschema/jsonschema/issues/1218

It also standardizes the message for `FILE_COPY_STATUS` notifications,
opting for the less verbose version.