Add bulk delete view for archive items (#3292)
Adds bulk delete view for archive items. I noticed that we have a custom
str method for archive items, so the confirmation view will look as
follows and won't contain the archive item pk. Should I update the str
method to also show the pk or remove the custom str method since we're
not using it anywhere else critical as far as I can tell?


![image](https://github.com/comic/grand-challenge.org/assets/30069334/037750ca-5f3a-4415-b7a1-cfee7706886c)