Add more information to job error messages (#3296)
Closes https://github.com/comic/grand-challenge.org/issues/2912

This adds the ValidationError messages to the job's error message and
adds a `detailed_error_message` field to the job model that we can use
to output the file import errors. The json and file validation errors
can go directly into the error message, as far as I can tell, since
these are short strings.