Update algorithm job detail API: add url and algorithm, remove algorithm_title (#3318)
Part of the pitch:
- https://github.com/DIAGNijmegen/rse-roadmap/issues/314

This PR adds a `url` to the job detail API to directly link to the
regular detail view. To be added to CIRRUS.

It also removes `algorithm_title` and replaces it with a hyperlinked
relation to the algorithm itself. The rationale is that the title has no
place in this serialisation. However, to get the title, a requester
would need to traverse via the `AlgorithmImage`, which only admins of an
algorithm would have access to (?). Adding a direct rel to the algorithm
enables a callee to retrieve the algorithm title via a secondary call to
get details on the algorithm. The algorithm details are public.