Display set titles (#3341)
Part of the pitch:
-
https://github.com/DIAGNijmegen/rse-roadmap/issues/313#event-12465816350

This adds a `title` field which is guaranteed to be unique for a display
set within a reader study.

In addition:
- Reader study copies now also take into account the ordering of display
sets
- Deleting list view of selected display set now shows the titles
- Negative display set ordering is now invalid on the form level

Some odd saves were going on concerning `order`. An initial version had
a `slug` as a unique counter to the `title`. This `slug` build-up
required the correct `order` at the time of saving. While the `slug` is
no longer needed in this version: I left the saving changes as I think
any data checking should have the correct order involved.

Once this PR passes the first inspection I'll throw out a new one with
similar changes for the archive items.


https://github.com/DIAGNijmegen/rse-roadmap/assets/7871310/8f96065c-1412-4955-8e7a-a1718ac6dceb