Add indicators for HTMX loading and POSTing (#3343)
This is a typical Friday afternoon fix.

This addresses a weird issue I had locally. I'd load the update page and
click save and nothing would happen...

By default, the form does a request. HTMX loading is deferred (something
which is good for speed), once HTMX passes over the DOM nodes the
behaviour of the buttons changes.

This PR disables the HTMX-related buttons until HTMX is initiated. It
adds an indicator while HTMX is loading and, for good measure, it now
also has an indicator to show HTMX is sending a request.

These attribute-driven changes can readily be added to future
interactions.

In addition, I've changed the titles of the Display set list to account
for tooltips not working on disabled UI elements.



https://github.com/comic/grand-challenge.org/assets/7871310/e6b953ad-5fb3-44c2-95cb-24a66a3ae2ff