Scoring: exclude_from_ranking (#3352)
Part of the tiny pitch:
- https://github.com/DIAGNijmegen/rse-roadmap/issues/322

This PR adds an optional (default: false) `exclude_from_ranking ` field
in the extra results columns JSON definition. Setting it to `true`
allows additional metrics to be shown on Challenge leaderboards while
being ignored while calculating the rank.

The related test is a construction copy of `test_calculate_ranks` since
I did not find it wise to extend that test with _yet_ another factor. It
would have pushed the 12 (2x3x2) to 24 which is over the top imho.

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>