Only allow linking algorithms with active image to archive (#3362)
At the moment, users can add algorithms without an active image to an
archive. When doing so, `create_algorithm_jobs_for_archive` will fail to
schedule jobs for those algorithms because of the missing image.

This PR catches the error and skips over algorithms without an active
image.
I also propose to limit the algorithm choices in the archive form to
those with active images.