Rework Component Interface Value representations (#3364)
Currently, we have several slightly different not-so-nice,
representations of component interface values (CIVs).

This PR contains a rework of these representations in the job detail
pages for evaluation and algorithms. The ultimate goal is to extend its
usage to the ArchiveItem detail page and ArchiveItem/DisplaySet list
pages.

### Current 'inline' view
<img width=50%
src="https://github.com/comic/grand-challenge.org/assets/7871310/4aaaac76-6323-4a46-858c-2be5e416095c">

### New 'inline' view
<img width=50%
src="https://github.com/comic/grand-challenge.org/assets/7871310/790bec0c-b671-4f1a-9f53-72f75a355c48">

## Template tags
~Two template tags are introduced: `civ` and `civ_inline`. Each tag
takes a component interface value as the only argument and renders an
HTML representation of the CIV. The inline variant is meant for
list-like representations where horizontal space is limited and relies
on a modal—the regular one where there is plenty of horizontal space to
render the CIV proper.~

A new template civ.html is introduced that accepts the arguments `civ`
and `display_inline`. The `display_inline` is a switch that can be set
to True and is meant for list-like representations where horizontal
space is limited and relies on a modal—the regular one where there is
plenty of horizontal space to render the CIV proper

## Vega Lazy loading
In addition, this PR adds lazy loading for vega charts which is very
much needed if viewing a list of CIVSets that output charts.

## New Fixtures
CIVs are a crucial part of development now, and I think adding a bunch
of test CIVs by hand is a lot of hassle. This, I think, plays a part in
these representations flying under the radar. As such, this PR also
introduces the `scripts/component_interface_value_fixtures.py`. This
repeatable script creates an algorithm job that has an overabundance of
different CIVs for quick development.

The screenshots are made based on this fixture.

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>