Apply lowering when cleaning Publication fields (#3400)
A minor issue popped up: validation on the form field did not do a
`lower()` on the identifier, meaning if any cases made it to the
publication (i.e. `"10.1016/S1470-2045(24)00220-1"`) it would not
validate.

It does seem that a DOI with case 'S' should be allowed, and since 's'
and 'S' are to be matched for duplication detection I've opted to clean
the field rather than change the regex.