Set the Celery soft time limit correctly (#3426)
The Celery soft time limit must always be shorter than the hard time
limit, otherwise the time limit is only enforced when the soft time
limit is reached. Celery will raise an exception on
`SoftTimeLimitExceeded` or `TimeLimitExceeded`, which will mark the task
as failed. We will have to monitor Celery tasks that fail with these
exceptions as in many places we were relying on the tasks to be
automatically rescheduled.

I've not tried this when using the explicit task arguments of
`time_limit` and `soft_time_limit`, but imagine they have the same
problem.

We cannot use `autoretry_on` as that makes tasks that occupy the
workers, not taking advantage of the delay queue.