Add custom celery task decorator (#3428)
This PR adds an opinionated decorator for celery tasks. It reduces the
possible options for a celery task. It forces us to use acks-late
queues, and adds common features for running tasks as singletons,
retries and error ignoring working around all of the bugs we have
encountered so far in Celery.