Fix failing rendering of archive item algorithm jobs page (#3450)
The issue was in a defined variable that was mistakingly put within
double quotes. That lead, in a downstream call, to raise an error when
attempting to access an attribute on that "string".

closes #3434 

The issue manifested in that datatables (JS dependency) was not able to
render a table of the jobs related to an archive item. The spinner
appears on the screen with an error message (browser alert) as shown
below:

![Screenshot 2024-07-17 at 20 48
02](https://github.com/user-attachments/assets/dc960ff5-30d0-4fe5-b5e8-bcd03276d57d)

After the fix, the table renders correctly, as shown below:

![Screenshot 2024-07-17 at 20 50
17](https://github.com/user-attachments/assets/d2ad074a-2581-4291-b7f5-42d11bfe144d)