Update UI running algorithm card to reflect average job duration (#3454)
Since running a job involves different stages other than executing the
algorithm on the selected image(s) (e.g. container pulling time,
provisioning and parsing times), the time viewed in the card is very
optimistic.

Therefore, it would be better (James suggestion) not to say how long is
remaining, rather how long a successful job has taken on average.

This small UI update PR changes the message in the algorithm running
card based on the status of the job. If the status is one of ("started",
"provisioning", "provisioned"), it will show a message that the job is
being provisioned.

![Screenshot 2024-07-25 at 09 41
00](https://github.com/user-attachments/assets/90372774-1a39-45b5-b8e6-904642eee4e9)


Otherwise, if the job has a status as one of ("executing", "executed",
"parsing outputs"), it will show a message reflecting the average job
duration and informing the user the the job is already being executed.

![Screenshot 2024-07-25 at 09 42
13](https://github.com/user-attachments/assets/eaf19552-3375-4307-8681-4608a704f474)


closes #3429