Update cost fixtures docker compose service (#3458)
The cost_fixtures task in Makefile uses the "web" service of the
docker-compose file which does not mount the docker socket to the
container. However, since this task requires building a docker image, it
won't be able to perform it from "web" and require the
"celery_worker_evaluation" service instead.