Add a check for evaluations with rank 0 and missing metrics and notify admin (#3459)
Implementing a solution to check for missing metrics and notify the
admin on the Evaluation detail page. Missing metrics lead to keep the
evaluation rank as zero and thus not showing in the leaderboard even
when it is finished with success and published.

A property is added to the Evaluation mode to check for missing metrics
in metrics.json. Moreover, a message is added to the details view
notifying the admin on which metrics are missing. Model and View tests
are provided for both functionalities.

![Screenshot 2024-07-29 at 11 44
48](https://github.com/user-attachments/assets/ed435792-5c32-4c73-9b7d-720942d0a165)



closes #3346