Fix failing algorithm usage chart to render (until resize) on Safari & Brave (#3460)
The original issue is reported on Safari browser (WebKit-based).
However, I had the same issue on Brave browser (Chromium-based), so it
spans different rendering engines. Similar issue was raised before
[here](https://talk.observablehq.com/t/embedding-vega-lite-charts-that-have-width-value-set-to-container-leads-to-null-width-until-window-is-resized/6120/5).

I found that on Algorithm details page, the page content is rendered (shown) through a a JQuery call on the active tab to show the content. On the other hand, the rendering of the chart happen in an event "DOMContentLoaded". At the time the chart rendering call is made, the content is not yet rendered on the screen by the JQuery call, and thus, the width: 'container'  of the chart is not properly working.

The solution is to register a listener for a custom event "vega.render" and then dispatch it after showing the tab for the first time in order to render the chart.