Add Challenge Registration Questions (#3526)
It introduces the creation, updating and deletion of registration
questions. It also has some validation on the level of models for when
answering is introduced.

To make things reviewable it does NOT contain the actual form changes
that will show the `RegistrationQuestion` and create the
`RegistrationQuestionAnswer` during sign-up to a challenge. It also does
not show the answers anywhere. That will be a follow-up PR.

See https://github.com/DIAGNijmegen/rse-roadmap/issues/344